+++ service/internal/telemetry/process_telemetry_test.go
@@ -24,9 +24,7 @@ import (
 )
 
 func TestProcessTelemetry(t *testing.T) {
-	const ballastSizeBytes uint64 = 0
-
-	pmv, err := NewProcessMetricsViews(ballastSizeBytes)
+	pmv, err := NewProcessMetricsViews()
 	require.NoError(t, err)
 	assert.NotNil(t, pmv)